backend/requirements.txt
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.21.0
sounddevice>=0.4.5
transformers>=4.30.0
accelerate>=0.20.0
bitsandbytes>=0.41.0
requests>=2.28.0
rich>=13.0.0 

flask
flask-cors
transformers
torch
sounddevice
numpy
python-dotenv
rich
groq